Add education policies after Easter (#66)
